almost finished implementation of single apk. the apk is generated correctly but i still have to change the package (totalcross.android), otherwise only a single totalcross application can be installed in a device. the classes.dex is being changed correctly, but i still have to compute the sha1 and crc32 to store at the front.

+++ TotalCrossSDK/src/tc/tools/deployer/Deployer4Android.java
@@ -9,8 +9,6 @@
  *                                                                               *
  *********************************************************************************/
 
-
-
 package tc.tools.deployer;
 
 import totalcross.util.*;
@@ -91,7 +89,10 @@ import java.util.zip.*;
 public class Deployer4Android
 {
    private static final boolean DEBUG = false;
-   private static String targetDir, targetPackage, jarOut, fileName;
+   private static String targetDir, sourcePackage, targetPackage, jarOut, fileName;
+   private String tcFolder = null, lbFolder = null;
+   private boolean isDemo,singleApk;
+
    byte[] buf = new byte[8192];
    
    public Deployer4Android() throws Exception
@@ -104,9 +105,31 @@ public class Deployer4Android
       File f = new File(targetDir);
       if (!f.exists())
          f.mkdirs();
+      singleApk = DeploySettings.packageType != 0;
+      if (!singleApk)
+         targetPackage = "totalcross/app/"+fileName.toLowerCase();
+      else
+      {
+         isDemo = (DeploySettings.packageType & DeploySettings.PACKAGE_DEMO) != 0;
+         tcFolder = (isDemo ? DeploySettings.folderTotalCrossSDKDistVM : DeploySettings.folderTotalCrossVMSDistVM)+"android/";
+         if ((DeploySettings.packageType & DeploySettings.PACKAGE_LITEBASE) != 0)
+            lbFolder = (isDemo ? DeploySettings.folderLitebaseSDKDistLIB : DeploySettings.folderLitebaseVMSDistLIB) + "android/";
+         sourcePackage = "totalcross/android";
+         targetPackage = "totalcross/"+fileName.toLowerCase();
+         // source and target packages must have the exact length
+         int slen = sourcePackage.length();
+         if (targetPackage.length() > slen)
+            targetPackage = targetPackage.substring(0,slen);
+         else
+            while (targetPackage.length() < slen)
+               targetPackage = targetPackage.concat("_");
+      }
       
-      createLauncher();  // 1
-      jar2dex();         // 2
+      if (!singleApk)
+      {
+         createLauncher();  // 1
+         jar2dex();         // 2
+      }
       updateResources(); // 3+4+5
       signAPK();         // 6
       
@@ -119,7 +142,6 @@ public class Deployer4Android
       if (jarIn == null)
          throw new DeployerException("File android/Launcher.jar not found!");
       jarOut = targetDir+fileName+".jar";
-      targetPackage = "totalcross/app/"+fileName.toLowerCase();
       
       ZipInputStream zis = new ZipInputStream(new FileInputStream(jarIn));
       ZipOutputStream zos = new ZipOutputStream(new FileOutputStream(jarOut));
@@ -183,16 +205,60 @@ public class Deployer4Android
          
          zos.closeEntry();
       }
-      // add classes.dex to the output zip file
-      zos.putNextEntry(new ZipEntry("classes.dex"));
-      totalcross.io.File f = new totalcross.io.File(targetDir+"classes.dex",totalcross.io.File.READ_WRITE);
-      int n;
-      while ((n=f.readBytes(buf,0,buf.length)) > 0)
-         zos.write(buf, 0, n);
-      f.delete(); // delete original file
+      if (singleApk)
+      {
+         processClassesDex(tcFolder+"TotalCross.apk", "classes.dex", zos, sourcePackage, targetPackage);
+         copyZipEntry(tcFolder+"TotalCross.apk", "res/layout/main.xml", zos);
+      }
+      else
+      {
+         // add classes.dex
+         zos.putNextEntry(new ZipEntry("classes.dex"));
+         totalcross.io.File f = new totalcross.io.File(targetDir+"classes.dex",totalcross.io.File.READ_WRITE);
+         int n;
+         while ((n=f.readBytes(buf,0,buf.length)) > 0)
+            zos.write(buf, 0, n);
+         zos.closeEntry();
+         f.delete(); // delete original file
+      }
+      // include the vm and litebase
+      if (tcFolder != null)
+         copyZipEntry(tcFolder+"TotalCross.apk", "lib/armeabi/libtcvm.so", zos);
+      if (lbFolder != null)
+         copyZipEntry(lbFolder+"Litebase.apk", "lib/armeabi/liblitebase.so", zos);
       
       zis.close();
       zos.close();      
+   }
+
+   private void processClassesDex(String srcZip, String fileName, ZipOutputStream dstZip, String from, String to) throws Exception
+   {
+      dstZip.putNextEntry(new ZipEntry(fileName));
+      byte[] bytes = Utils.loadZipEntry(srcZip,fileName);
+      if (from != null) // replace all strings "from" to "to"
+      {
+         int ofs=0;
+         byte[] fromBytes = from.getBytes();
+         byte[] toBytes = to.getBytes();
+         while ((ofs = Utils.indexOf(bytes, fromBytes, false, ofs)) != -1)
+         {
+            totalcross.sys.Vm.arrayCopy(toBytes, 0, bytes, ofs, toBytes.length);
+            ofs += toBytes.length;
+         }
+      }
+      // FALTA RECALCULAR O CHECKSUM E O SHA-1
+      // 0x8   4  Checksum
+      // 0xC   20    SHA-1 Signature
+      dstZip.write(bytes,0,bytes.length);
+      dstZip.closeEntry();
+   }
+   
+   private void copyZipEntry(String srcZip, String fileName, ZipOutputStream dstZip) throws Exception
+   {
+      dstZip.putNextEntry(new ZipEntry(fileName));
+      byte[] bytes = Utils.loadZipEntry(srcZip,fileName);
+      dstZip.write(bytes,0,bytes.length);
+      dstZip.closeEntry();
    }
    
    private void signAPK() throws Exception
@@ -233,28 +299,39 @@ public class Deployer4Android
       return bas;
    }
    
-   private void insertAndroidManifest_xml(ZipInputStream zis, ZipOutputStream zos) throws Exception
+   private void insertAndroidManifest_xml(InputStream zis, OutputStream zos) throws Exception
    {
-      totalcross.io.ByteArrayStream bas = readInputStream(zis);
+      totalcross.io.ByteArrayStream bas;
+      if (!singleApk)
+         bas = readInputStream(zis);
+      else
+      {
+         byte[] bytes = Utils.loadZipEntry(DeploySettings.etcDir+"tools/android/single.apk","AndroidManifest.xml");
+         bas = new totalcross.io.ByteArrayStream(bytes);
+         bas.skipBytes(bytes.length);
+      }
       bas.mark();
       totalcross.io.DataStreamLE ds = new totalcross.io.DataStreamLE(bas);
       
-      String oldPackage  = "totalcross.app.stub";
+      String oldPackage  = singleApk ? sourcePackage.replace('/','.') : "totalcross.app.stub";
       String oldVersion  = "1.0";
       String oldTitle    = "Stub";
-      String oldActivity = ".Stub";
+      String oldActivity = singleApk ? null : ".Stub";
+      String oldSharedId = singleApk ? "totalcross.app.sharedid" : null;
       
       String newPackage  = targetPackage.replace('/','.');
       String newVersion  = DeploySettings.appVersion != null ? DeploySettings.appVersion : oldVersion;
       String newTitle    = DeploySettings.appTitle;
-      String newActivity = "."+fileName;
+      String newActivity = singleApk ? null : "."+fileName;
+      String newSharedId = singleApk ? "totalcross.app."+fileName.toLowerCase() : null;
       
       int oldSize = bas.available();
       int difPackage  = (newPackage .length() - oldPackage .length()) * 2;
       int difVersion  = (newVersion .length() - oldVersion .length()) * 2;
       int difTitle    = (newTitle   .length() - oldTitle   .length()) * 2;
-      int difActivity = (newActivity.length() - oldActivity.length()) * 2;
-      int dif = difPackage + difVersion + difTitle + difActivity;
+      int difActivity = singleApk ? 0 : (newActivity.length() - oldActivity.length()) * 2;
+      int difSharedId = !singleApk ? 0 : (newSharedId.length() - oldSharedId.length()) * 2;
+      int dif = difPackage + difVersion + difTitle + difActivity + difSharedId;
       
       // get the xml size
       bas.setPos(12); 
@@ -292,16 +369,20 @@ public class Deployer4Android
       int resSize = bas.available();
       byte[] res = new byte[resSize];
       bas.readBytes(res,0,resSize);
-      // search and replace the value of versionCode="305419896" (0x12345678) with the current date and time in format YYMMDDHHMM (year with 2 digits only)
-      byte[] versionCodeMark = {(byte)0x78,(byte)0x56,(byte)0x34,(byte)0x12};
-      int ofs = Utils.indexOf(res, versionCodeMark, false);
-      if (ofs == -1)
-         throw new DeployerException("Error: could not find position for versionCode");
-      int datetime = getDateTime();
-      totalcross.io.ByteArrayStream dtbas = new totalcross.io.ByteArrayStream(res);
-      dtbas.setPos(ofs);
-      totalcross.io.DataStreamLE dsbas = new totalcross.io.DataStreamLE(dtbas);
-      dsbas.writeInt(datetime);
+      int ofs;
+      if (false) // this approach is no longer used.
+      {
+         // search and replace the value of versionCode="305419896" (0x12345678) with the current date and time in format YYMMDDHHMM (year with 2 digits only)
+         byte[] versionCodeMark = {(byte)0x78,(byte)0x56,(byte)0x34,(byte)0x12};
+         ofs = Utils.indexOf(res, versionCodeMark, false);
+         if (ofs == -1)
+            throw new DeployerException("Error: could not find position for versionCode");
+         int datetime = getDateTime();
+         totalcross.io.ByteArrayStream dtbas = new totalcross.io.ByteArrayStream(res);
+         dtbas.setPos(ofs);
+         totalcross.io.DataStreamLE dsbas = new totalcross.io.DataStreamLE(dtbas);
+         dsbas.writeInt(datetime);
+      }
       // if is full screen, search and replace Theme.Black.NoTitleBar by Theme.Black.NoTitleBar.Fullscreen
       if (DeploySettings.isFullScreenPlatform(totalcross.sys.Settings.ANDROID)) // guich@tc120_59
       {
@@ -313,20 +394,23 @@ public class Deployer4Android
       }
       
       // now, change the names accordingly
-      for (int i =0; i < len; i++)
+      for (int i = 0; i < len; i++)
       {
          String s = strings[i];
-         if (s.equals(oldPackage))
+         if (oldPackage != null && s.equals(oldPackage))
             strings[i] = newPackage;
          else
-         if (s.equals(oldVersion))
+         if (oldVersion != null && s.equals(oldVersion))
             strings[i] = newVersion;
          else
-         if (s.equals(oldTitle))
+         if (oldTitle != null && s.equals(oldTitle))
             strings[i] = newTitle;
          else
-         if (s.equals(oldActivity))
+         if (oldActivity != null && s.equals(oldActivity))
             strings[i] = newActivity;
+         else
+         if (oldSharedId != null && s.equals(oldSharedId))
+            strings[i] = newSharedId;
       }
       // update the offsets table
       for (int i = 0; i < len; i++)
@@ -365,10 +449,20 @@ public class Deployer4Android
       return year * 100000000 + month * 1000000 + day * 10000 + hour * 100 + minute;
    }
 
-   private void insertResources_arsc(ZipInputStream zis, ZipOutputStream zos) throws Exception
+   private void insertResources_arsc(InputStream zis, OutputStream zos) throws Exception
    {
-      byte[] key = {'t',(byte)0,'o',(byte)0,'t',(byte)0,'a',(byte)0,'l',(byte)0,'c',(byte)0,'r',(byte)0,'o',(byte)0,'s',(byte)0,'s',(byte)0,'.',(byte)0,'a',(byte)0,'p',(byte)0,'p',(byte)0,'.',(byte)0,'s',(byte)0,'t',(byte)0,'u',(byte)0,'b',(byte)0};
-      byte[] all = readInputStream(zis).toByteArray();
+      byte[] all;
+      byte[] key;
+      if (singleApk)
+      {
+         key = new byte[]{'t',(byte)0,'o',(byte)0,'t',(byte)0,'a',(byte)0,'l',(byte)0,'c',(byte)0,'r',(byte)0,'o',(byte)0,'s',(byte)0,'s',(byte)0,'.',(byte)0,'a',(byte)0,'n',(byte)0,'d',(byte)0,'r',(byte)0,'o',(byte)0,'i',(byte)0,'d',(byte)0};
+         all = Utils.loadZipEntry(DeploySettings.etcDir+"tools/android/single.apk","resources.arsc");
+      }
+      else
+      {
+         key = new byte[]{'t',(byte)0,'o',(byte)0,'t',(byte)0,'a',(byte)0,'l',(byte)0,'c',(byte)0,'r',(byte)0,'o',(byte)0,'s',(byte)0,'s',(byte)0,'.',(byte)0,'a',(byte)0,'p',(byte)0,'p',(byte)0,'.',(byte)0,'s',(byte)0,'t',(byte)0,'u',(byte)0,'b',(byte)0};
+         all = readInputStream(zis).toByteArray();
+      }
       int ofs = Utils.indexOf(all, key, false);
       if (ofs == -1)
          throw new DeployerException("Could not find position for totalcross.app.stub in arsc.");
@@ -405,6 +499,15 @@ public class Deployer4Android
       vLocals.addElement(DeploySettings.tczFileName);
       if (vGlobals.size() > 0)
          vLocals.addElements(vGlobals.toObjectArray());
+      if (singleApk) // include the vm?
+      {
+         // tc is always included
+         // include non-binary files
+         vLocals.addElement(DeploySettings.folderTotalCrossSDKDistVM+"TCBase.tcz");
+         vLocals.addElement(DeploySettings.folderTotalCrossSDKDistVM+"TCFont.tcz");
+         if ((DeploySettings.packageType & DeploySettings.PACKAGE_LITEBASE) != 0)
+            vLocals.addElement(DeploySettings.folderLitebaseSDKDistLIB+"LitebaseLib.tcz");
+      }         
 
       for (int i =0, n = vLocals.size(); i < n; i++)
       {
@@ -417,6 +520,9 @@ public class Deployer4Android
             if (name.startsWith("/"))
                name = name.substring(1);
          }         
+         // tcz's name must match the lowercase sharedid
+         if (tcFolder != null && pathname.equals(DeploySettings.tczFileName)) 
+            name = name.toLowerCase();
          zos.putNextEntry(new ZipEntry(name));
          FileInputStream fis = new FileInputStream(pathname);
          copyStream(fis, zos);
